Fixed Multiplication function yet again
Bug Details - The multiplication function was giving seg fault for multiplying by any number with zeros, this was because of the add function which was returning the original list if the passed list was null. But then when changing the temp list in multiply now the original list was being changed, which was resulting in error in the calculation part of the multiplication.    FIX - Updated the add functionions opening condition to create a new list and copy the whole non-empty list to it in O(n) time and then return the newly allocated list